Move entity temporal surface tests into entities package (#293)
