+++ internal/graph/entities/facets.go
@@ -1,0 +1,23 @@
+package entities
+
+import graph "github.com/writer/cerebro/internal/graph"
+
+func GetEntityFacetDefinition(id string) (EntityFacetDefinition, bool) {
+	return graph.GetEntityFacetDefinition(id)
+}
+
+func DefaultEntityFacetDefinitions() []EntityFacetDefinition {
+	return graph.ListEntityFacetDefinitions()
+}
+
+func EntityFacetAppliesToNode(def EntityFacetDefinition, kind graph.NodeKind) bool {
+	if len(def.ApplicableKinds) == 0 {
+		return true
+	}
+	for _, candidate := range def.ApplicableKinds {
+		if candidate == kind {
+			return true
+		}
+	}
+	return false
+}
